[backend] Code refactoring
