+++ backend/src/main/java/ca/gov/dtsstn/cdcp/api/Application.java
@@ -10,9 +10,6 @@ import org.springframework.context.event.EventListener;
 
 import ca.gov.dtsstn.cdcp.api.config.properties.ApplicationProperties;
 
-/**
- * @author Greg Baker <gregory.j.baker@hrsdc-rdhcc.gc.ca>
- */
 @SpringBootApplication
 @EnableConfigurationProperties({ ApplicationProperties.class })
 public class Application {
