--- backend/src/main/java/ca/gov/dtsstn/cdcp/api/ApplicationProperties.java
@@ -1,15 +1,0 @@
-package ca.gov.dtsstn.cdcp.api;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.validation.annotation.Validated;
-
-/**
- * @author Greg Baker <gregory.j.baker@hrsdc-rdhcc.gc.ca>
- */
-@Validated
-@ConfigurationProperties("application")
-public class ApplicationProperties {
-
-	/* placeholder for future configuration */
-
-}
