+++ backend/src/main/java/ca/gov/dtsstn/cdcp/api/web/v1/controller/SubscriptionController.java
@@ -1,6 +1,5 @@
-package ca.gov.dtsstn.cdcp.api.controller;
+package ca.gov.dtsstn.cdcp.api.web.v1.controller;
 
-import java.util.Collection;
 import java.util.List;
 
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,15 +7,15 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import ca.gov.dtsstn.cdcp.api.dto.ImmutableSubscription;
-import ca.gov.dtsstn.cdcp.api.dto.Subscription;
+import ca.gov.dtsstn.cdcp.api.web.v1.model.ImmutableSubscription;
+import ca.gov.dtsstn.cdcp.api.web.v1.model.Subscription;
 
 @RestController
 @RequestMapping({ "/api/v1" })
-public class SubscriptionControllerV1 {
+public class SubscriptionController {
 
 	@GetMapping({ "/users/{userId}/subscriptions" })
-	public Collection<Subscription> getSubscriptions(@PathVariable String userId) {
+	public List<Subscription> getSubscriptions(@PathVariable String userId) {
 		return List.of(
 			ImmutableSubscription.builder()
 				.id(10001L)
